Guard pack generation against an Int-overflow OOM
`needed = packCount * packSize` was an Int multiply. The web
/cube/api/generate endpoint puts no upper bound on packs/packSize, so a
crafted request (e.g. packs=2^30, packSize=4) overflowed `needed` to 0,
slipped past the "not enough cards" guard, and hit List(packCount) — a
~1-billion-element allocation (OOM / DoS).

Compute `needed` as Long so oversized requests correctly fail the pool
check with a clean error (a 400 on the web). Anything that passes the
check has needed ≤ pool size, so it still fits an Int for the draw.
Discord was already safe (its options cap at 90 packs / 30 size).

Test: a 2^30 × 4 request returns a "not enough cards" Failure fast
instead of OOMing, plus an Int.MAX_VALUE pack size.

+++ common/src/main/kotlin/common/mtg/PackGenerator.kt
@@ -47,7 +47,12 @@ class PackGenerator(private val random: Random = Random.Default) {
     ): Result {
         if (packCount <= 0) return Result.Failure("Pack count must be at least 1 (was $packCount).")
         if (packSize <= 0) return Result.Failure("Pack size must be at least 1 (was $packSize).")
-        val needed = packCount * packSize
+        // Long multiply: packCount × packSize can overflow Int (e.g. a crafted
+        // web request with packs=2^30, packSize=4 wraps to 0 and would slip past
+        // this guard into a huge List(packCount) allocation — an OOM). With the
+        // pool check below, anything that passes has needed ≤ pool.size, so it
+        // safely fits back in an Int.
+        val needed = packCount.toLong() * packSize.toLong()
         if (pool.size < needed) {
             return Result.Failure(
                 "Not enough cards: need $needed ($packCount × $packSize) but the pool only has ${pool.size}."
@@ -55,7 +60,7 @@ class PackGenerator(private val random: Random = Random.Default) {
         }
 
         // Randomly select the exact number of cards we'll deal.
-        val selected = pool.shuffled(random).take(needed)
+        val selected = pool.shuffled(random).take(needed.toInt())
 
         // The stream we deal from. Balanced → cards grouped into contiguous
         // category blocks (still shuffled within each block, via the stable
